Generalise viewmodel classes across the project
